+++ crates/tad-gui/src/compiler_thread.rs
@@ -15,6 +15,7 @@ use crate::audio_thread::{
 use std::collections::hash_map::Entry;
 use std::collections::HashMap;
 use std::fmt::Display;
+use std::ops::Deref;
 use std::sync::{mpsc, Arc};
 use std::thread;
 
@@ -38,7 +39,7 @@ use compiler::songs::{
 };
 use compiler::sound_effects::{
     blank_compiled_sound_effects, combine_sound_effects, tad_gui_sfx_data,
-    CombinedSoundEffectsData, CompiledSfxMap,
+    CombinedSoundEffectsData, CompiledSfxMap, SfxExportOrder,
 };
 use compiler::sound_effects::{compile_sound_effect_input, CompiledSoundEffect, SoundEffectInput};
 use compiler::spc_file_export::export_spc_file;
@@ -90,6 +91,7 @@ impl<T> ReplaceAllVec<T> {
 #[derive(Debug)]
 pub struct ProjectToCompiler {
     pub sfx_export_order: ReplaceAllVec<data::Name>,
+    pub low_priority_sfx_export_order: ReplaceAllVec<data::Name>,
     pub pf_songs: ReplaceAllVec<data::Song>,
     pub instruments: ReplaceAllVec<data::Instrument>,
     pub samples: ReplaceAllVec<data::Sample>,
@@ -110,6 +112,8 @@ pub enum ToCompiler {
     ClearSampleCacheAndRebuild,
 
     SfxExportOrder(ItemChanged<data::Name>),
+    LowPrioritySfxExportOrder(ItemChanged<data::Name>),
+
     ProjectSongs(ItemChanged<data::Song>),
 
     Instrument(ItemChanged<data::Instrument>),
@@ -181,25 +185,28 @@ impl SfxId {
 }
 
 #[derive(Debug)]
-pub struct SfxExportOrder {
-    pub names: Vec<data::Name>,
+pub struct GuiSfxExportOrder {
+    export_order: Vec<data::Name>,
+    low_priority_index: usize,
 }
 
-impl SfxExportOrder {
-    pub fn new(names: Vec<data::Name>) -> SfxExportOrder {
-        SfxExportOrder { names }
+impl SfxExportOrder for GuiSfxExportOrder {
+    fn n_sound_effects(&self) -> usize {
+        self.export_order.len()
     }
 
-    pub fn n_sound_effects(&self) -> usize {
-        self.names.len()
+    fn export_order(&self) -> &[data::Name] {
+        &self.export_order
     }
 
-    pub fn export_order(&self) -> &[data::Name] {
-        &self.names
+    fn low_priority_index(&self) -> usize {
+        self.low_priority_index
     }
+}
 
+impl GuiSfxExportOrder {
     pub fn sfx_id(&self, index: usize) -> Option<SfxId> {
-        if index < self.names.len() {
+        if index < self.export_order.len() {
             index.try_into().ok().map(SfxId)
         } else {
             None
@@ -210,7 +217,7 @@ impl SfxExportOrder {
 #[derive(Debug)]
 pub struct CommonAudioDataWithSfx {
     pub common_audio_data: CommonAudioData,
-    pub sfx_export_order: Arc<SfxExportOrder>,
+    pub sfx_export_order: Arc<GuiSfxExportOrder>,
 }
 
 #[derive(Debug)]
@@ -719,12 +726,19 @@ fn create_sfx_compiler<'a>(
     }
 }
 
-fn build_sfx_export_order(eo: &IList<data::Name>) -> Arc<SfxExportOrder> {
-    Arc::new(SfxExportOrder::new(eo.items().to_vec()))
+fn build_sfx_export_order(
+    eo: &IList<data::Name>,
+    lp_eo: &IList<data::Name>,
+) -> Arc<GuiSfxExportOrder> {
+    // Assumes eo and lp_eo do not contain duplicate names
+    Arc::new(GuiSfxExportOrder {
+        export_order: [eo.items(), lp_eo.items()].concat(),
+        low_priority_index: eo.items().len(),
+    })
 }
 
 fn count_missing_sfx(
-    sfx_export_order: &SfxExportOrder,
+    sfx_export_order: &GuiSfxExportOrder,
     sound_effects: &CList<SoundEffectInput, Option<Arc<CompiledSoundEffect>>>,
     sender: &Sender,
 ) {
@@ -744,7 +758,7 @@ fn count_missing_sfx(
 }
 
 fn calc_sfx_data_size(
-    sfx_export_order: &SfxExportOrder,
+    sfx_export_order: &GuiSfxExportOrder,
     sound_effects: &CList<SoundEffectInput, Option<Arc<CompiledSoundEffect>>>,
 ) -> usize {
     let sfx_size: usize = sfx_export_order
@@ -776,7 +790,7 @@ impl SongDependencies {
 fn build_common_data_no_sfx_and_song_dependencies(
     instruments: &CList<data::Instrument, Option<InstrumentSampleData>>,
     samples: &CList<data::Sample, Option<SampleSampleData>>,
-    sfx_export_order: &SfxExportOrder,
+    sfx_export_order: &GuiSfxExportOrder,
     sound_effects: &CList<SoundEffectInput, Option<Arc<CompiledSoundEffect>>>,
     blank_sfx_data: &CombinedSoundEffectsData,
 ) -> Result<(Arc<CommonAudioDataWithSfxBuffer>, SongDependencies), CombineSamplesError> {
@@ -832,22 +846,22 @@ impl CompiledSfxMap for CList<SoundEffectInput, Option<Arc<CompiledSoundEffect>>
 /// Returns `(CommonAudioDataWithSize, sfx_data_size)`
 fn build_common_audio_data_with_sfx(
     dep: &Option<SongDependencies>,
-    sfx_export_order: Arc<SfxExportOrder>,
+    sfx_export_order: &Arc<GuiSfxExportOrder>,
     sound_effects: &CList<SoundEffectInput, Option<Arc<CompiledSoundEffect>>>,
 ) -> (Option<Arc<CommonAudioDataWithSfx>>, usize) {
-    let sfx_data = combine_sound_effects(sound_effects, sfx_export_order.export_order());
+    let sfx_data = combine_sound_effects(sound_effects, sfx_export_order.deref());
 
     // Always try to calculate sfx_data size (to keep song size checks and Project tab up-to-date)
     let sfx_data_size = match &sfx_data {
         Ok(s) => s.sfx_data_size(),
-        Err(_) => calc_sfx_data_size(&sfx_export_order, sound_effects),
+        Err(_) => calc_sfx_data_size(sfx_export_order, sound_effects),
     };
 
     let cad = match (&dep, &sfx_data) {
         (Some(dep), Ok(sd)) => match build_common_audio_data(&dep.combined_samples, sd) {
             Ok(cad) => Some(Arc::new(CommonAudioDataWithSfx {
                 common_audio_data: cad,
-                sfx_export_order,
+                sfx_export_order: sfx_export_order.clone(),
             })),
             Err(_) => None,
         },
@@ -1185,7 +1199,8 @@ fn bg_thread(
 
     let blank_sfx_data = tad_gui_sfx_data(SFX_BUFFER_SIZE);
 
-    let mut sfx_export_order_ilist = IList::new();
+    let mut sfx_eo_ilist = IList::new();
+    let mut lp_sfx_eo_ilist = IList::new();
     let mut pf_songs = IList::new();
     let mut instruments = CList::new();
     let mut samples = CList::new();
@@ -1199,7 +1214,7 @@ fn bg_thread(
 
     let mut inst_sample_names = Arc::default();
 
-    let mut sfx_export_order = build_sfx_export_order(&sfx_export_order_ilist);
+    let mut sfx_export_order = build_sfx_export_order(&sfx_eo_ilist, &lp_sfx_eo_ilist);
 
     // Used to test if a sound effect was edited in the sfx tab.
     let mut cad_with_sfx_out_of_date = true;
@@ -1214,12 +1229,13 @@ fn bg_thread(
             ToCompiler::LoadProject(p) => {
                 songs.replace_all_and_load_songs(&p.pf_songs);
 
-                sfx_export_order_ilist.replace_all(p.sfx_export_order);
+                sfx_eo_ilist.replace_all(p.sfx_export_order);
+                lp_sfx_eo_ilist.replace_all(p.low_priority_sfx_export_order);
                 pf_songs.replace_all(p.pf_songs);
                 instruments.replace_all(p.instruments);
                 samples.replace_all(p.samples);
 
-                sfx_export_order = build_sfx_export_order(&sfx_export_order_ilist);
+                sfx_export_order = build_sfx_export_order(&sfx_eo_ilist, &lp_sfx_eo_ilist);
                 inst_sample_names = instrument_and_sample_names(&instruments, &samples);
 
                 compile_all_samples(
@@ -1251,8 +1267,15 @@ fn bg_thread(
             }
 
             ToCompiler::SfxExportOrder(m) => {
-                sfx_export_order_ilist.process_message(m);
-                sfx_export_order = build_sfx_export_order(&sfx_export_order_ilist);
+                sfx_eo_ilist.process_message(m);
+                sfx_export_order = build_sfx_export_order(&sfx_eo_ilist, &lp_sfx_eo_ilist);
+                count_missing_sfx(&sfx_export_order, &sound_effects, &sender);
+
+                pending_build_cad_with_sfx = true;
+            }
+            ToCompiler::LowPrioritySfxExportOrder(m) => {
+                lp_sfx_eo_ilist.process_message(m);
+                sfx_export_order = build_sfx_export_order(&sfx_eo_ilist, &lp_sfx_eo_ilist);
                 count_missing_sfx(&sfx_export_order, &sound_effects, &sender);
 
                 pending_build_cad_with_sfx = true;
@@ -1447,7 +1470,7 @@ fn bg_thread(
 
             let (cad, sfx_data_size) = build_common_audio_data_with_sfx(
                 &song_dependencies,
-                sfx_export_order.clone(),
+                &sfx_export_order,
                 &sound_effects,
             );